compile antlr4k module also for browser and node
